docs: moved provider getResourcesToSpend docs snippets
