+++ apps/docs-snippets2/src/provider/functionality-get-resources-to-spend.ts
@@ -1,0 +1,46 @@
+// #region getResourcesToSpend-1
+import type { CoinQuantityLike, ExcludeResourcesOption } from 'fuels';
+import { Provider, ScriptTransactionRequest, Wallet } from 'fuels';
+
+import { LOCAL_NETWORK_URL, WALLET_PVT_KEY } from '../env';
+
+const provider = await Provider.create(LOCAL_NETWORK_URL);
+const wallet = Wallet.fromPrivateKey(WALLET_PVT_KEY);
+const assetIdA =
+  '0x0101010101010101010101010101010101010101010101010101010101010101';
+
+const baseAssetId = provider.getBaseAssetId();
+
+const quantities: CoinQuantityLike[] = [
+  { amount: 32, assetId: baseAssetId, max: 42 },
+  { amount: 50, assetId: assetIdA },
+];
+
+const utxoId =
+  '0x00000000000000000000000000000000000000000000000000000000000000010001';
+const messageNonce =
+  '0x381de90750098776c71544527fd253412908dec3d07ce9a7367bd1ba975908a0';
+const excludedIds: ExcludeResourcesOption = {
+  utxos: [utxoId],
+  messages: [messageNonce],
+};
+
+const spendableResources = await provider.getResourcesToSpend(
+  wallet.address,
+  quantities,
+  excludedIds
+);
+
+const tx = new ScriptTransactionRequest();
+tx.addResources(spendableResources);
+// #endregion getResourcesToSpend-1
+
+// #region getResourcesToSpend-2
+const spendableResourcesFromWallet = await wallet.getResourcesToSpend(
+  spendableResources,
+  excludedIds
+);
+// #endregion getResourcesToSpend-2
+
+console.log(spendableResources);
+console.log(spendableResourcesFromWallet);
